requirements.txt
aiogram == 2.22.1
aioschedule == 0.5.2
beautifulsoup4 == 4.11.1
requests == 2.28.0
mysql-connector-python == 8.0.31
python-dotenv == 0.21.0
